+++ src/main/java/lotto/model/LottoNumbersIssuance.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
