+++ src/main/java/lotto/model/Person.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import lotto.views.ResultView;
