+++ src/main/java/lotto/model/PlaceStatus.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 public enum PlaceStatus {
     FIRST, SECOND, THIRD, FOURTH, FIFTH, NOTHING
