+++ src/main/java/lotto/model/Result.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import java.util.List;
